LPS-83949 Move servlets to their proper package
